+++ spec/rubocop/cop/rspec/variable_name_spec.rb
@@ -188,4 +188,24 @@ RSpec.describe RuboCop::Cop::RSpec::VariableName, :config do
       end
     end
   end
+
+  context 'when configured to ignore certain patterns' do
+    let(:cop_config) do
+      { 'EnforcedStyle' => 'snake_case',
+        'IgnoredPatterns' => ['^userFood$', '^userPet$'] }
+    end
+
+    it 'registers an offense when not matching any ignored patterns' do
+      expect_offense(<<~RUBY)
+        let(:userName) { 'Adam' }
+            ^^^^^^^^^ Use snake_case for variable names.
+      RUBY
+    end
+
+    it 'does not register an offense when matching any ignored pattern' do
+      expect_no_offenses(<<~RUBY)
+        let(:userFood) { 'Adam' }
+      RUBY
+    end
+  end
 end